Move files, cleanup
