+++ Core/Core/Features/Inbox/InboxSettings/View/InboxSettingsView.swift
@@ -95,8 +95,7 @@ public struct InboxSettingsView: View {
                     .accessibilityHidden(true)
                     .padding(.vertical, defaultPadding)
 
-                UITextViewWrapper(text: $viewModel.signature) {
-                    let tv = UITextView()
+                InstUI.UITextViewWrapper(text: $viewModel.signature) { tv in
                     tv.placeholder = String(localized: "Write your signature here", bundle: .core)
                     tv.placeholderColor = .textPlaceholder
                     tv.isScrollEnabled = false
@@ -106,7 +105,6 @@ public struct InboxSettingsView: View {
                     tv.translatesAutoresizingMaskIntoConstraints = false
                     tv.widthAnchor.constraint(equalToConstant: geometry.frame(in: .global).width - (2 * defaultPadding)).isActive = true
                     tv.backgroundColor = .backgroundLightest
-                    return tv
                 }
                 .font(.regular16, lineHeight: .condensed)
                 .textInputAutocapitalization(.sentences)
