+++ Teacher/Teacher/SpeedGrader/Comments/View/OldCommentEditorView.swift
@@ -31,8 +31,12 @@ struct OldCommentEditorView: View {
 
     var body: some View {
         HStack(alignment: .bottom) {
-            DynamicHeightTextEditor(text: $text, placeholder: String(localized: "Comment", bundle: .teacher), font: .regular16)
-                .lineLimit(10)
+            InstUI.ScrollableTextEditor(
+                text: $text,
+                placeholder: String(localized: "Comment", bundle: .teacher),
+                font: .regular16,
+                lineLimit: 10
+            )
                 .accessibility(label: Text("Comment", bundle: .teacher))
                 .identifier("SubmissionComments.commentTextView")
                 .highPriorityGesture( // High priority to take precedence over comment field activation.
